pyproject.toml
[project]
name = "presumption-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.68.1",
    "datasets>=4.1.1",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]
